move joystick register reading code from core back into joystick
